钱包及账变提交

+++ wallet-modules-site-api/src/main/java/com/ruoyi/wallet/api/dto/BalanceChangeDTO.java
@@ -42,7 +42,7 @@ public class BalanceChangeDTO implements Serializable {
      * 账变类型
      */
     @NotNull
-    private Integer businessId;
+    private Long businessId;
     /**
      * 币种
      */
